moved Challenge specific classes
